requirements.txt
streamlit
PyMuPDF
python-docx
pillow
pytesseract
cohere
sentence-transformers
torch